Move transaction validator out of tx pool component
